Feature/teleportpoint2d (#20)
* 'update'

* wallet import? & start on general player character and world

* outside world camera and boundaries

- added an outside world scene that allows the user to walk around
- added world boundaries keeps user inside
- added camera boundaries that limits camera inside the map
- limited player space until further features are developed

* outside world camera and boundaries

- added an outside world scene that allows the user to walk around
- added world boundaries keeps user inside
- added camera boundaries that limits camera inside the map
- limited player space until further features are developed
- updated tile set up in order set up future teleport room

* TeleportPoint2D initial commit

still working a little on figuring out the best way. And now I found it with the help of AI. I will be updating this and building a better solution for it later

* teleports fucntionality

created teleports that can be placed and linked to, additionally updated project files, added a scene manger and updates some shop and player stuff

* semi stable teleport system

teleports are added in all rooms. All rooms have been connected and play can go in between them.